requirements.txt
# requirements.txt
Django>=4.2,<5.0
gunicorn>=21.2
whitenoise>=6.5
python-dotenv>=1.0.0